make names make more sense with how they ended up being used
