--- packages/nodes/src/views/document-view/components/context-menu-container/index.ts
@@ -1,3 +1,0 @@
-import ContextMenuContainer from './ContextMenuContainer'
-
-export { ContextMenuContainer }
