--- packages/nodes/src/views/document-view/components/grid-container/index.ts
@@ -1,3 +1,0 @@
-import GridContainer from './GridContainer'
-
-export { GridContainer }
