--- packages/nodes/src/views/document-view/components/index.ts
@@ -1,4 +1,0 @@
-export { CameraOverlay } from './camera-overlay'
-export { ContextMenuContainer } from './context-menu-container'
-export { GridContainer } from './grid-container'
-export { NodePlacementOverlay } from './node-placement-overlay'
