+++ packages/nodes/src/views/document-view/layers/camera-overlay/CameraOverlay.tsx
