+++ packages/nodes/src/views/document-view/layers/canvas-grid-underlay/index.ts
@@ -1,0 +1,3 @@
+import CanvasGridUnderlay from './CanvasGridUnderlay'
+
+export { CanvasGridUnderlay }
