+++ packages/nodes/src/views/document-view/layers/index.ts
@@ -1,0 +1,4 @@
+export { CameraOverlay } from './camera-overlay'
+export { CanvasGridUnderlay } from './canvas-grid-underlay'
+export { NodePlacementOverlay } from './node-placement-overlay'
+export { TransientElementOverlay } from './transient-element-overlay'
