+++ packages/nodes/src/views/document-view/layers/node-placement-overlay/NodePlacementOverlay.tsx
