+++ packages/nodes/src/views/document-view/layers/transient-element-overlay/index.ts
@@ -1,0 +1,3 @@
+import TransientElementOverlay from './TransientElementOverlay'
+
+export { TransientElementOverlay }
